commitment: move test code to own package
To avoid circular dependencies between the test code and upcoming
changes to the mock code, we make sure that the tests actually run in a
different Go package.
We only need to do that for test code that uses any mocks that we're
going to move to their own package later on.