multi: move gRPC related code out of fn package
To support compiling certain code with Tinygo, we need to be careful
that we don't use external dependencies that are incompatible.
Because the error helper functions pull in the gRPC package, we move
that code to the taprpc package that won't be part of the binary we
compile with Tinygo.