tapchannel: add new DeriveTapscriptRoot method
This matches the recent changes on the lnd side. We can't return the full funding desc until we know the contributions from the remote party. However, by the time we're ready to send open channel, we already know the tapscript root (just uses an OP_TRUE output), so we can return that to make sure lnd knows the correct funding output.

This covers the change for the initiator. The changes for the responder will be in the next commit, as we add a new `asset_funding_created` message.