tapchannel: update pre funding flow for input+output asset verification
In this commit, we update the pre funding flow for the responder. At the end of the pre funding flow, the responder now has the TAP asset commitment that'll be used in the funding output.

With this, once they receive the OpenChannel message, they can now obtain the tapscript root needed to properly create any musig2 sessions and also the funding output.

For the initiator, we'll now also send the output proof to the responder within sendInputOwnershipProofs.